gaze_overlay_video: now check if source file has audio, and only attempt to add audio track if it does

+++ src/glassesTools/gaze_overlay_video.py
@@ -3,6 +3,7 @@ import cv2
 import pathlib
 import shutil
 import os
+import subprocess
 from enum import Enum, auto
 
 from ffpyplayer.writer import MediaWriter
@@ -128,29 +129,40 @@ class VideoMaker:
 
     def finish_video(self):
         self._vid_writer.close()
-        # if ffmpeg is on path, add audio to scene and optionally board video
-        if shutil.which('ffmpeg') is not None:
-            # move file to temp name
-            tempName = self.output_path.with_stem(self.output_path.stem + '_temp')
-            shutil.move(self.output_path, tempName)
+        # if ffmpeg is on path, add audio to scene video
+        if (shutil.which('ffmpeg') is not None) and (shutil.which('ffprobe') is not None):
+            # check if source file has audio
+            command = ['ffprobe',
+               '-loglevel', 'error',
+               '-select_streams', 'a',
+               '-show_entries', 'stream=codec_type',
+               '-of', 'csv=p=0',
+               f'{self.src_video}']
+            proc = subprocess.Popen(command, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
+            out, err = proc.communicate()
+            if not err and out.decode().strip()=='audio':
+                # file has audio, lets go
+                # move file to temp name
+                tempName = self.output_path.with_stem(self.output_path.stem + '_temp')
+                shutil.move(self.output_path, tempName)
 
-            # add audio
-            cmds = ['ffmpeg', '-hide_banner', '-loglevel', 'error', '-y', '-i', f'"{tempName}"', '-i', f'"{self.src_video}"', '-vcodec', 'copy']
-            # determine if can copy audio or need to transcode
-            # we transcode if source file is not mp4-style, to be safe. Some audio formats that are allowed in
-            # avi and mkv are not allowed in mp4
-            if video_utils.is_isobmmf(self.src_video):
-                # copy. else, specifying nothing, it will be transcoded
-                cmds.extend(['-acodec', 'copy'])
-            cmds.extend(['-map', '0:v:0', '-map', '1:a:0?', '-shortest', f'"{self.output_path}"'])
-            os.system(' '.join(cmds))
+                # add audio
+                cmds = ['ffmpeg', '-hide_banner', '-loglevel', 'error', '-y', '-i', f'"{tempName}"', '-i', f'"{self.src_video}"', '-vcodec', 'copy']
+                # determine if can copy audio or need to transcode
+                # we transcode if source file is not mp4-style, to be safe. Some audio formats that are allowed in
+                # avi and mkv are not allowed in mp4
+                if video_utils.is_isobmmf(self.src_video):
+                    # copy. else, specifying nothing, it will be transcoded
+                    cmds.extend(['-acodec', 'copy'])
+                cmds.extend(['-map', '0:v:0', '-map', '1:a:0?', '-shortest', f'"{self.output_path}"'])
+                os.system(' '.join(cmds))
 
-            # clean up
-            if self.output_path.exists():
-                tempName.unlink(missing_ok=True)
-            else:
-                # something failed. Put file without audio back under output name
-                shutil.move(tempName, self.output_path)
+                # clean up
+                if self.output_path.exists():
+                    tempName.unlink(missing_ok=True)
+                else:
+                    # something failed. Put file without audio back under output name
+                    shutil.move(tempName, self.output_path)
 
     def process_video(self):
         while True:
